+++ backend/src/main/java/com/grepp/smartwatcha/app/controller/api/recommend/payload/MovieRecommendUserBasedResponse.java
@@ -3,7 +3,6 @@ package com.grepp.smartwatcha.app.controller.api.recommend.payload;
 import com.grepp.smartwatcha.infra.jpa.entity.MovieEntity;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
-
 import java.time.LocalDateTime;
 import java.util.List;
