+++ backend/src/main/java/com/grepp/smartwatcha/app/controller/web/IndexController.java
@@ -9,12 +9,11 @@ import com.grepp.smartwatcha.app.model.index.dto.IndexMovieDto;
 import com.grepp.smartwatcha.app.model.notification.NotificationService;
 import com.grepp.smartwatcha.app.model.recommend.RecommendPersonalMovieService;
 import com.grepp.smartwatcha.app.model.recommend.RecommendUserBasedMovieService;
-import com.grepp.smartwatcha.app.model.recommend.service.RecommendHighestRatedJpaService;
+import com.grepp.smartwatcha.app.model.recommend.service.highestrated.RecommendHighestRatedJpaService;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
-import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
