+++ backend/src/main/java/com/grepp/smartwatcha/app/model/recommend/repository/MovieGenreNeo4jRepository.java
@@ -6,7 +6,9 @@ import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
 
+
 @Repository
 public interface MovieGenreNeo4jRepository extends Neo4jRepository<MovieNode, Long> {
+    // neo4j 내부 id로 노드 조회
     Optional<MovieNode> findById(Long id);
 }
