+++ backend/src/main/java/com/grepp/smartwatcha/app/model/recommend/repository/MovieQueryJpaRepository.java
@@ -1,33 +1,21 @@
 package com.grepp.smartwatcha.app.model.recommend.repository;
 
 import com.grepp.smartwatcha.infra.jpa.entity.MovieEntity;
-import jakarta.persistence.EntityManager;
-import jakarta.persistence.PersistenceContext;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
 
 @Repository
-public class MovieQueryJpaRepository {
+public interface MovieQueryJpaRepository extends JpaRepository<MovieEntity, Long> {
 
-    @PersistenceContext
-    private EntityManager em;
+    // 공개된 영화 목록을 조회함
+    @Query("SELECT m FROM MovieEntity m WHERE m.isReleased = true")
+    List<MovieEntity> findAllReleased();
 
-    public MovieEntity findById(Long id) {
-        return em.find(MovieEntity.class, id);
-    }
-
-    public List<MovieEntity> findAllReleased() {
-        return em.createQuery(
-                        "SELECT m FROM MovieEntity m WHERE m.isReleased = true", MovieEntity.class)
-                .getResultList();
-    }
-
-    public List<MovieEntity> findByIdIn(List<Long> ids) {
-        return em.createQuery("""
-                SELECT m FROM MovieEntity m WHERE m.id IN :ids
-            """, MovieEntity.class)
-                .setParameter("ids", ids)
-                .getResultList();
-    }
+    // 여러개의 id를 조회
+    @Query("SELECT m FROM MovieEntity m WHERE m.id IN :ids")
+    List<MovieEntity> findByIdIn(@Param("ids") List<Long> ids);
 }
